backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.82.0",
    "fastapi>=0.135.2",
    "pgvector>=0.3.6",
    "psycopg[binary,pool]>=3.3.3",
    "python-multipart>=0.0.22",
    "sentence-transformers>=3.0.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "uvicorn>=0.42.0",
    "torch",
    "torchvision",
    "torchaudio",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true  # Das hier ist extrem wichtig!

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }
